refactor(review): share the call wording
The navigator is about to print the same call changes the details bar
does, so the wording moves to `labels::calls` where both can read it.
`text::call_lines` and friends stay as re-exports, so the bar is
untouched. `signature_pair` is new: one line for a column too narrow
for the two-line token diff.
